refactor: use general paths for imports
