+++ src/api/logToken.ts
@@ -1,5 +1,5 @@
 import type { APIResponse, Login, RequestOptions, Token } from "..";
-import { apiRequest, formatDate } from "../util";
+import { apiRequest, formatDate } from "..";
 
 /**
  * Log the token.
